Move throttle to picard.util.qt
throttle relies on QTimer, so it moves to the Qt-only picard.util.qt.
Update the importers (ui.metadatabox, ui.tablebaseddialog).
